+++ src/simmate/visualization/structure/__init__.py
@@ -1,1 +1,1 @@
-from .blender.base import make_blender_structure
+from .blender import make_blender_structure